refactor cache and cli
