+++ packages/relative-values/src/components/Interface/ModelPicker.tsx
@@ -9,7 +9,8 @@ import { useSelectedInterface } from "./InterfaceProvider";
 
 const ModelPickerMenu: FC<{ close(): void }> = ({ close }) => {
   const { models, catalog } = useSelectedInterface();
-  const { selectedId } = useSelectedModel();
+  const selectedModel = useSelectedModel();
+  const selectedId = selectedModel?.id;
 
   const router = useRouter();
 
@@ -58,7 +59,7 @@ const ModelPickerMenu: FC<{ close(): void }> = ({ close }) => {
 };
 
 export const ModelPicker: FC = () => {
-  const { selectedModel } = useSelectedModel();
+  const selectedModel = useSelectedModel();
 
   return (
     <Dropdown render={({ close }) => <ModelPickerMenu close={close} />}>
