+++ packages/relative-values/src/components/View/hooks/useRelativeValues.ts
@@ -1,19 +1,20 @@
-import { getModelCode, Model } from "@/model/utils";
+import { Model } from "@/model/utils";
+import { jsonData } from "@/values/cache";
 import { RVStorage } from "@/values/RVStorage";
-import { jsonData } from "@/values/RVCache";
 
-export const useRelativeValues = (
-  id: string | undefined,
-  model: Model | undefined
-) => {
+// TODO: doesn't have to be a hook, and changing it into a static method on RVStorage would help with `model: undefined` case
+export const useRelativeValues = (model: Model | undefined) => {
   let cache = jsonData
     .flatMap((r) => r.models)
     .find((r) => model && r.id == model.id);
+  if (!model) {
+    return { error: "No model specified" };
+  }
   if (!cache) {
     return { error: "Data not in cache" };
   }
   return {
     error: "",
-    rv: new RVStorage(cache),
+    rv: new RVStorage(model, cache),
   };
 };
